Allow preferences sheets to be enabled/disabled
Allow for control over whether preferences sheets are
accepted for a given show. When preferences sheets are closed,
the audition numbers for that show are generated.

Display text on the show page if preferences sheets are enabled,
inviting a logged in user to submit a sheet (and prompting a user
to log in, if not). Preferences sheets are unavailable when closed.

* Add column to show model (requires schema change)
* Add messages to show_detail to advertise the show's preferences
  sheets, when they are open
* Restrict preferences sheets to only shows that have them open
* Automatically generate audition numbers when closing preferences
  sheets for a show.